requirements.txt
pandas
pykrx
python-dotenv
